move common table service code
